requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
youtube-transcript-api
